+++ src/collective/iconifiedcategory/behaviors/iconifiedcategorization.py
@@ -16,11 +16,10 @@ from zope.component import adapter
 from zope.interface import Interface
 from zope.interface import implementer
 from zope.interface import provider
-from collective.z3cform.select2.widget.widget import SingleSelect2FieldWidget
 
 from collective.iconifiedcategory import _
-from collective.iconifiedcategory import utils
 from collective.iconifiedcategory.widget.widget import CategoryTitleFieldWidget
+from collective.z3cform.select2.widget.widget import SingleSelect2FieldWidget
 
 
 @provider(IFormFieldProvider)
@@ -61,19 +60,6 @@ class IconifiedCategorization(object):
 
     @content_category.setter
     def content_category(self, value):
-        if self.content_category is None:
-            category = utils.get_category_object(self.context, value)
-            # left False if print/confidential
-            # not enabled on ContentCategoryGroup
-            category_group = category.get_category_group(category)
-            if category_group.to_be_printed_activated:
-                self.context.to_print = category.to_print
-            else:
-                self.context.to_print = False
-            if category_group.confidentiality_activated:
-                self.context.confidential = category.confidential
-            else:
-                self.context.confidential = False
         self.context.content_category = value
         self.context.reindexObject(idxs=['content_category_uid'])
 
@@ -84,3 +70,11 @@ class IconifiedCategorization(object):
     @property
     def confidential(self):
         return getattr(aq_base(self.context), 'confidential', False)
+
+    @property
+    def to_sign(self):
+        return getattr(aq_base(self.context), 'to_sign', False)
+
+    @property
+    def signed(self):
+        return getattr(aq_base(self.context), 'signed', False)