+++ CHANGES.rst
@@ -4,8 +4,19 @@ Changelog
 0.23 (unreleased)
 -----------------
 
-- Nothing changed yet.
-
+- Added management of `to_sign` and `signed` attributes like it is the case for
+  `to_print` and `confidential` attributes.  Both attributes are used behind a
+  single action `signed` that have 3 options : `not to sign`, `to sign` and
+  `signed`.
+  [gbastien]
+- Default values for `to_print`, `confidential` and `to_sign/signed` are now
+  managed in the `IObjectAddedEvent` no more in the `content_category setter`,
+  this way every attribtues are managed the same way because `to_print` and
+  `confidential` are simple attributes where `to_sign/signed` can come from the
+  `Scan metadata` behavior of `collective.dms.scanbehavior`.
+  [gbastien]
+- Added possibility to show/hide details about `to_print`, `confidential`,
+  `to_sign/signed` in the categorized elements tooltipster.
 
 0.22 (2017-08-04)
 -----------------